AI/agno-agent/requirements.txt
agno
openai
python-dotenv
crawl4ai
playwright
beautifulsoup4
requests
pandas
mcp
fastapi
uvicorn